refactor browserless config to remote module
