+++ dendrite_sdk/sync_api/_ext_impl/browserless/_impl.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, Optional
 from dendrite_sdk._common._exceptions.dendrite_exception import BrowserNotLaunchedError
 from dendrite_sdk.sync_api._core._impl_browser import ImplBrowser
 from dendrite_sdk.sync_api._core._type_spec import PlaywrightPage
-from dendrite_sdk.sync_api._ext_impl.browserless._settings import BrowserlessConfig
+from dendrite_sdk.remote.browserless_config import BrowserlessConfig
 
 if TYPE_CHECKING:
     from dendrite_sdk.sync_api._core.dendrite_browser import Dendrite
